Busy Signal and Multi-Threading
This commit provides a better user experience by showing a progress bar or busy signal while the image is being processed. The image is processed in a different thread than the dynamic tab creation and busy bar. Once the image has finished processing, the busy signal stops and the image is shown.